Vulcan Viz - move reorder piece to own file
